fix signedness for proto bounding circle radius

+++ src/tyr/locate_serializer.cc
@@ -151,7 +151,7 @@ void serialize_edges(const PathLocation& path_location,
         writer.start_object("bounding_circle");
         writer("lat", edge.bounding_circle.first.lat());
         writer("lon", edge.bounding_circle.first.lng());
-        writer("radius", static_cast<int64_t>(edge.bounding_circle.second));
+        writer("radius", edge.bounding_circle.second);
         writer.end_object();
 
         writer("linear_reference", linear_reference(directed_edge, edge.percent_along, edge_info));
